requirements.txt
python-dotenv
python-telegram-bot
requests
flask
pyyaml
tweepy
deta